feat(snapRemote) create a provider w/ snap options¬
Add the ability to define global snap options into angular config phase.
The main goal is to define global options in a global area and not necessarily in a controller. The latter solution still works and overwrite options defined in the config phase. It makes sense to do so, i.e. in "resolver" context or runtime modifications.

There is no breaking change.
3 unit tests added, good code coverage.
Some tests updated.